feat: remove GradusBase
Removed GradusBase from within the code. This is then effectively a
refactor to move code around, and rename a handful of files.

Also removed the DocStringExtensions dependency, since it's not really
that useful at this stage of writing documentation.
