--- docs/src/api-documentation/GradusBase.md
@@ -1,6 +1,0 @@
-# GradusBase API documentation
-
-```@autodocs
-Modules = [Gradus.GradusBase]
-Order   = [:function, :type]
-```
