+++ src/geometry/discs.jl
@@ -51,8 +51,6 @@ r_cross_section(d::AbstractThickAccretionDisc, r::Number) =
     struct GeometricThinDisc{T} <: AbstractAccretionDisc{T}
     GeometricThinDisc(inner_radius::T, outer_radius::T, inclination::T)
 
-$(FIELDS)
-
 Simple geometrically thin accretion disc spanning from `inner_radius` to `outer_radius` in
 gravitational units. Inclination of the disc is relative to spin axis, with ``90^\\circ``
 being perpendicular to the spin axis.
