+++ src/geometry/meshes.jl
@@ -1,7 +1,5 @@
 """
     MeshAccretionGeometry(mesh)
-
-$(FIELDS)
 """
 struct MeshAccretionGeometry{T} <: AbstractAccretionGeometry{T}
     mesh::Vector{Tuple{SVector{3,T},SVector{3,T},SVector{3,T}}}
