+++ src/metrics/johannsen-ad.jl
@@ -39,7 +39,12 @@ end # module
     struct JohannsenMetric{T} <: AbstractStaticAxisSymmetric{T}
 
 The Johannsen (20xx) metric.
-$(FIELDS)
+- `M = 1.0`: Black hole mass.
+- `a = 0.0`: Black hole spin.
+- `α13 = 0.0`: ``\\alpha_{13}`` deviation parameter.
+- `α22 = 0.0`: ``\\alpha_{22}`` deviation parameter.
+- `α52 = 0.0`: ``\\alpha_{52}`` deviation parameter.
+- `ϵ3 = 0.0`: ``\\epsilon_{3}`` deviation parameter.
 """
 @with_kw struct JohannsenMetric{T} <: AbstractStaticAxisSymmetric{T}
     @deftype T
@@ -59,6 +64,6 @@ end
 
 metric_components(m::JohannsenMetric{T}, rθ) where {T} =
     __JohannsenAD.metric_components(m, rθ)
-GradusBase.inner_radius(m::JohannsenMetric{T}) where {T} = m.M + √(m.M^2 - m.a^2)
+inner_radius(m::JohannsenMetric{T}) where {T} = m.M + √(m.M^2 - m.a^2)
 
 export JohannsenMetric
