+++ src/utils.jl
@@ -32,4 +32,43 @@ end
 
 cartesian_distance(m::AbstractMetric, x1, x2) = √(cartesian_squared_distance(m, x1, x2))
 
+# both energy and angular momentum
+# assume time only coupled to radial coordinate
+# need to think of a nice way to keep this efficient
+# whilst allowing metrics with other couplings
+
+"""
+    E(m::AbstractMatrix{T}, v)
+    E(m::AbstractMetric{T}, u, v)
+
+Compute the energy for a numerically evaluated metric, and some velocity four vector `v`,
+```math
+E = - p_t = - g_{t\\nu} p^\\nu.
+```
+
+For null geodesics, the velocity is the momentum ``v^\\nu = p^\\nu``. For massive geodesics,
+the mass ``\\mu`` needs to be known to compute ``\\mu v^\\nu = p^\\nu``.
+"""
+function E(metric::AbstractMatrix{T}, v) where {T}
+    T(@inbounds -(metric[1, 1] * v[1] + metric[1, 4] * v[4]))
+end
+E(m::AbstractMetric, u, v) = E(metric(m, u), v)
+E(m::AbstractMetric, gp::AbstractGeodesicPoint) = E(m, gp.x, gp.v)
+
+
+"""
+    Lz(m::AbstractMatrix{T}, v)
+    Lz(m::AbstractMetric{T}, u, v)
+
+Compute the angular momentum for a numerically evaluated metric, and some velocity four vector `v`.
+```math
+L_z = p_\\phi = - g_{\\phi\\nu} p^\\nu.
+```
+"""
+function Lz(metric::AbstractMatrix{T}, v) where {T}
+    T(@inbounds metric[4, 4] * v[4] + metric[1, 4] * v[1])
+end
+Lz(m::AbstractMetric, u, v) = Lz(metric(m, u), v)
+Lz(m::AbstractMetric, gp::AbstractGeodesicPoint) = Lz(m, gp.x, gp.v)
+
 export cartesian_squared_distance, cartesian_distance, spherical_to_cartesian
